diff to _diff and consolidate resource to abstract class'
