+++ mica-rest/src/main/java/org/obiba/mica/dataset/rest/collection/DraftCollectedDatasetResource.java
@@ -10,10 +10,8 @@
 
 package org.obiba.mica.dataset.rest.collection;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.MapDifference;
 
 import org.obiba.mica.AbstractGitPersistableResource;
 import org.obiba.mica.JSONUtils;
@@ -25,7 +23,6 @@ import org.obiba.mica.dataset.domain.StudyDataset;
 import org.obiba.mica.dataset.domain.StudyDatasetState;
 import org.obiba.mica.dataset.service.CollectedDatasetService;
 import org.obiba.mica.security.rest.SubjectAclResource;
-import org.obiba.mica.study.service.DocumentDifferenceService;
 import org.obiba.mica.web.model.Dtos;
 import org.obiba.mica.web.model.Mica;
 import org.obiba.opal.web.model.Magma;
@@ -181,30 +178,6 @@ public class DraftCollectedDatasetResource extends
   public Mica.DatasetDto getFromCommit(@NotNull @PathParam("commitId") String commitId) throws IOException {
     checkPermission("/draft/collected-dataset", "VIEW");
     return dtos.asDto(datasetService.getFromCommit(datasetService.findDraft(id), commitId), true);
-  }
-
-  @GET
-  @Path("/diff")
-  public Response diff(@NotNull @QueryParam("left") String left, @NotNull @QueryParam("right") String right) {
-    checkPermission("/draft/individual-study", "VIEW");
-
-    StudyDataset leftCommit = datasetService.getFromCommit(datasetService.findDraft(id), left);
-    StudyDataset rightCommit = datasetService.getFromCommit(datasetService.findDraft(id), right);
-
-    Map<String, Object> data = new HashMap<>();
-
-    try {
-      MapDifference<String, Object> difference = DocumentDifferenceService.diff(leftCommit, rightCommit);
-
-      data.put("differing", DocumentDifferenceService.fromEntriesDifferenceMap(difference.entriesDiffering()));
-      data.put("onlyLeft", difference.entriesOnlyOnLeft());
-      data.put("onlyRight", difference.entriesOnlyOnRight());
-
-    } catch (JsonProcessingException e) {
-      //
-    }
-
-    return Response.ok(data, MediaType.APPLICATION_JSON_TYPE).build();
   }
 
   @Path("/permissions")
