+++ mica-rest/src/main/java/org/obiba/mica/dataset/rest/harmonization/DraftHarmonizedDatasetResource.java
@@ -11,10 +11,8 @@
 package org.obiba.mica.dataset.rest.harmonization;
 
 import com.codahale.metrics.annotation.Timed;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.MapDifference;
 
 import org.obiba.mica.AbstractGitPersistableResource;
 import org.obiba.mica.JSONUtils;
@@ -27,7 +25,6 @@ import org.obiba.mica.dataset.domain.HarmonizationDataset;
 import org.obiba.mica.dataset.domain.HarmonizationDatasetState;
 import org.obiba.mica.dataset.service.HarmonizedDatasetService;
 import org.obiba.mica.security.rest.SubjectAclResource;
-import org.obiba.mica.study.service.DocumentDifferenceService;
 import org.obiba.mica.web.model.Dtos;
 import org.obiba.mica.web.model.Mica;
 import org.obiba.opal.web.model.Magma;
@@ -202,30 +199,6 @@ public class DraftHarmonizedDatasetResource extends
   public Mica.DatasetDto getFromCommit(@NotNull @PathParam("commitId") String commitId) throws IOException {
     checkPermission("/draft/harmonized-dataset", "VIEW");
     return dtos.asDto(datasetService.getFromCommit(datasetService.findDraft(id), commitId), true);
-  }
-
-  @GET
-  @Path("/diff")
-  public Response diff(@NotNull @QueryParam("left") String left, @NotNull @QueryParam("right") String right) {
-    checkPermission("/draft/individual-study", "VIEW");
-
-    HarmonizationDataset leftCommit = datasetService.getFromCommit(datasetService.findDraft(id), left);
-    HarmonizationDataset rightCommit = datasetService.getFromCommit(datasetService.findDraft(id), right);
-
-    Map<String, Object> data = new HashMap<>();
-
-    try {
-      MapDifference<String, Object> difference = DocumentDifferenceService.diff(leftCommit, rightCommit);
-
-      data.put("differing", DocumentDifferenceService.fromEntriesDifferenceMap(difference.entriesDiffering()));
-      data.put("onlyLeft", difference.entriesOnlyOnLeft());
-      data.put("onlyRight", difference.entriesOnlyOnRight());
-
-    } catch (JsonProcessingException e) {
-      //
-    }
-
-    return Response.ok(data, MediaType.APPLICATION_JSON_TYPE).build();
   }
 
   @Path("/permissions")
