+++ mica-rest/src/main/java/org/obiba/mica/study/rest/DraftHarmonizationStudyResource.java
@@ -14,7 +14,6 @@ import com.codahale.metrics.annotation.Timed;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
-import com.google.common.collect.MapDifference;
 import com.google.common.collect.Maps;
 import org.obiba.mica.AbstractGitPersistableResource;
 import org.obiba.mica.JSONUtils;
@@ -31,7 +30,6 @@ import org.obiba.mica.security.rest.SubjectAclResource;
 import org.obiba.mica.study.ConstraintException;
 import org.obiba.mica.study.domain.HarmonizationStudy;
 import org.obiba.mica.study.domain.HarmonizationStudyState;
-import org.obiba.mica.study.service.DocumentDifferenceService;
 import org.obiba.mica.study.service.HarmonizationStudyService;
 import org.obiba.mica.web.model.Dtos;
 import org.obiba.mica.web.model.Mica;
@@ -206,30 +204,6 @@ public class DraftHarmonizationStudyResource extends AbstractGitPersistableResou
   public Mica.StudyDto getStudyFromCommit(@NotNull @PathParam("commitId") String commitId) throws IOException {
     checkPermission("/draft/harmonization-study", "VIEW");
     return dtos.asDto(studyService.getFromCommit(studyService.findDraft(id), commitId), true);
-  }
-
-  @GET
-  @Path("/diff")
-  public Response diff(@NotNull @QueryParam("left") String left, @NotNull @QueryParam("right") String right) {
-    checkPermission("/draft/individual-study", "VIEW");
-
-    HarmonizationStudy leftCommit = studyService.getFromCommit(studyService.findDraft(id), left);
-    HarmonizationStudy rightCommit = studyService.getFromCommit(studyService.findDraft(id), right);
-
-    Map<String, Object> data = new HashMap<>();
-
-    try {
-      MapDifference<String, Object> difference = DocumentDifferenceService.diff(leftCommit, rightCommit);
-
-      data.put("differing", DocumentDifferenceService.fromEntriesDifferenceMap(difference.entriesDiffering()));
-      data.put("onlyLeft", difference.entriesOnlyOnLeft());
-      data.put("onlyRight", difference.entriesOnlyOnRight());
-
-    } catch (JsonProcessingException e) {
-      //
-    }
-
-    return Response.ok(data, MediaType.APPLICATION_JSON_TYPE).build();
   }
 
   @Path("/permissions")
